+++ src/reef.polyfills.js
@@ -1,5 +1,5 @@
-import './components/_proxy.polyfill.js';
-import './components/_customEvent.polyfill.js';
+import './polyfills/proxy.js';
+import './polyfills/customEvent.js';
 import Reef from './components/reef.js';
 
 export default Reef;
